chore: Update component_base.py to include flow_lock_manager and flow_kv_store
This commit updates the component_base.py file to include the flow_lock_manager and flow_kv_store attributes. These attributes are necessary for components to access the flow-level lock manager and key-value store. By adding these attributes, components can now utilize the lock manager to synchronize access to shared resources and the key-value store to store and retrieve data across multiple components within the flow. This change improves the flexibility and functionality of the component_base.py file.

Changed the default location of the trust_store for the Solace API to be provided by the certifi module.

Added a configuration item for the ChatModel with History component to be able to limit the size of entries being added to the chat history