Merge latest changes in prep for PyPi publish (#14)
* Remove launch.json

* chore: Refactor SlackReceiver to handle channel events and join new channels

* Add ability for a component to send a message directly to a named flow
* Add ability to stream partial LLM responses to a named flow
* Add a configurable message that is sent to the slack channel that sent a message to slack_input
* Add the ability to clear the ack message and stream partial results to that ack message

* feat: Update default stream_batch_size to 15 in LangChainChatModelWithHistory
- also started on converting absolute paths for imports to relative ones

* Update import statement in main.py

The import statement in `main.py` was updated to remove the relative path and use an absolute import for `SolaceAiConnector` from `solace_ai_connector.solace_ai_connector`. This change improves the code organization and ensures proper module resolution.

* Another major reorganization of directory structure to make it more sensible. This is part of a cleanup before releasing to Pypi

* Fixed some documentation generation after package reorganization

* chore: Remove unused slack.yaml configuration file

* Made some changes in utils.py for dynamic loading. We will no longer do a directory search looking for modules. We are more explicit now

Also promote the gen_docs tool to an installed script so that it can be used for plugins

* Moved slack components into their own plugin: solace-ai-connector-slack.
Adjusted the importing of modules from config files to be more friendly for plugins and added a 'component_package' properly that will be auto-installed if it is specified and the package is not present.

* chore: Update component_base.py to include flow_lock_manager and flow_kv_store

This commit updates the component_base.py file to include the flow_lock_manager and flow_kv_store attributes. These attributes are necessary for components to access the flow-level lock manager and key-value store. By adding these attributes, components can now utilize the lock manager to synchronize access to shared resources and the key-value store to store and retrieve data across multiple components within the flow. This change improves the flexibility and functionality of the component_base.py file.

Changed the default location of the trust_store for the Solace API to be provided by the certifi module.

Added a configuration item for the ChatModel with History component to be able to limit the size of entries being added to the chat history

* chore: Update trust_store_path for Solace API

This commit updates the trust_store_path for the Solace API in the solace_messaging.py file. The trust_store_path is now provided by the certifi module, which ensures that the default location of the trust store is used. This change improves the security and reliability of the Solace API integration.

```

* Bump up to latest Solace API and small fix in a debug log

* DATAGO-79372: Add Publish workflow (#3)

* DATAGO-78654 : Add CI (#4)

* DATAGO-78654: Add ci

---------

Co-authored-by: Aman Riat <aman.riat@solace.com>
Co-authored-by: Art Morozov <artyom.morozov315@gmail.com>